+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
-- Correct conversion of nba play by play shot distance from int32 to float32
+- Correct conversion of nba play by play shot distance from int32 to float32 (#120)
 
 ## [0.18.0] - 2026-02-07
 ### Added